Module hyperion_kernel::syscall

Structs

File 🔒

ProcessExtra 🔒

Functions

call_id 🔒

close

close a file

exit

exit and kill the current process

log

print a string to logs

map_vfs_err_to_syscall_err 🔒

nanosleep

sleep at least arg0 nanoseconds

nanosleep_until

sleep at least until the nanosecond arg0 happens

open

open a file

palloc

allocate physical pages and map them to virtual memory

pfree

free allocated physical pages

process_ext_with 🔒

pthread_spawn

spawn a new thread

read

read bytes from a file

read_slice_parts 🔒

read_untrusted_bytes 🔒

read_untrusted_bytes_mut 🔒

read_untrusted_str 🔒

recv

recv data from this process input channel

rename

rename the current process

send

send data to an input channel of a process

syscall

timestamp

get the number of nanoseconds after boot

write

write bytes into a file

yield_now

give the processor back to the kernel temporarily